% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_ffd.R
\name{anova_ffd}
\alias{anova_ffd}
\title{Calcula ANOVA do modelo fatorial}
\usage{
anova_ffd(modelo)
}
\arguments{
\item{modelo}{Objeto retornado por fit_ffd()}
}
\value{
Tabela de ANOVA
}
\description{
Calcula ANOVA do modelo fatorial
}
\examples{
# anova_ffd(fit)
}
