% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaliar_ponto_estacionario_ffd.R
\name{avaliar_ponto_estacionario_ffd}
\alias{avaliar_ponto_estacionario_ffd}
\title{Avalia o ponto estacionário em relação ao objetivo}
\usage{
avaliar_ponto_estacionario_ffd(modelo, objetivo = c("min", "max"))
}
\arguments{
\item{modelo}{Objeto retornado por fit_ffd().}

\item{objetivo}{"min" para minimizar ou "max" para maximizar.}
}
\value{
Lista com avaliação interpretativa do ponto estacionário.
}
\description{
Avalia o ponto estacionário em relação ao objetivo
}
