% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeficientes_ffd.R
\name{coeficientes_ffd}
\alias{coeficientes_ffd}
\title{Extrai coeficientes do modelo fatorial}
\usage{
coeficientes_ffd(modelo)
}
\arguments{
\item{modelo}{Objeto retornado por fit_ffd()}
}
\value{
Data frame com os coeficientes do modelo
}
\description{
Extrai coeficientes do modelo fatorial
}
\examples{
# coeficientes_ffd(fit)
}
