% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contorno_ffd.R
\name{contorno_ffd}
\alias{contorno_ffd}
\title{Gráfico de contorno para FFD com ponto ótimo e estacionário}
\usage{
contorno_ffd(
  modelo,
  x1,
  x2,
  n = 60,
  mostrar_otimo = TRUE,
  mostrar_estacionario = TRUE,
  objetivo = "min"
)
}
\arguments{
\item{modelo}{Objeto retornado por fit_ffd().}

\item{x1}{Nome do primeiro fator.}

\item{x2}{Nome do segundo fator.}

\item{n}{Número de pontos da grade.}

\item{mostrar_otimo}{Exibir ponto ótimo.}

\item{mostrar_estacionario}{Exibir ponto estacionário.}

\item{objetivo}{"min" ou "max".}
}
\description{
Gráfico de contorno para FFD com ponto ótimo e estacionário
}
