% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efeitos_ffd.R
\name{efeitos_ffd}
\alias{efeitos_ffd}
\title{Calcula efeitos do modelo fatorial}
\usage{
efeitos_ffd(modelo, alpha = 0.05)
}
\arguments{
\item{modelo}{Objeto retornado por fit_ffd()}

\item{alpha}{Nível de significância (default = 0.05)}
}
\value{
Data frame com efeitos, tipo e significância
}
\description{
Calcula efeitos do modelo fatorial
}
\examples{
# efeitos_ffd(fit)
}
