% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar_matriz_ffd.R
\name{exportar_matriz_ffd}
\alias{exportar_matriz_ffd}
\title{Exporta matriz fatorial para Excel}
\usage{
exportar_matriz_ffd(matriz, caminho = NULL)
}
\arguments{
\item{matriz}{Data frame gerado por matriz_ffd()}

\item{caminho}{Caminho para salvar o arquivo (opcional)}
}
\description{
Exporta matriz fatorial para Excel
}
\examples{
m <- matriz_ffd(fatores = c("A","B"), niveis = 2)
exportar_matriz_ffd(m)
}
