%
\name{FFD-internal}
\alias{computeSampleSizeInd}
\alias{ltdSampling.internal}
\alias{indSampling.internal}
\alias{computeDiseaseMatrix}
\alias{roundConstantSum}
\alias{createStyleFile}
\alias{importDataFun}
\alias{chooseHerdSizeCol}
\alias{onOkHerdSize}
\alias{saveText}
\alias{savePlot}
\alias{buildMySurvey}
\alias{OnCalcSS}
\alias{OnSampleSS}
\alias{sampleFarms}
\alias{FFDmanual}
\alias{saveFFD}
\alias{loadFFD}
\alias{loadExample}
\alias{resetFFD}
\alias{plotWidget}
\alias{OnCalcDiag}
\title{Internal FFD functions}
\description{
  Internal FFD functions, methods and classes.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
