% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ffd.R
\name{fit_ffd}
\alias{fit_ffd}
\title{Ajusta modelo para planejamento fatorial completo}
\usage{
fit_ffd(dados, resposta, fatores)
}
\arguments{
\item{dados}{Data frame com os dados experimentais.}

\item{resposta}{Nome da variável resposta.}

\item{fatores}{Vetor com os nomes dos fatores.}
}
\value{
Objeto lm com o modelo ajustado.
}
\description{
Ajusta modelo para planejamento fatorial completo
}
