% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriz_ffd().R
\name{matriz_ffd}
\alias{matriz_ffd}
\title{Gera matriz de planejamento fatorial completo}
\usage{
matriz_ffd(fatores, niveis = 2, replicas = 1, exportar = FALSE, caminho = NULL)
}
\arguments{
\item{fatores}{Vetor com os nomes dos fatores}

\item{niveis}{Número de níveis (2 ou 3)}

\item{replicas}{Número de réplicas (default = 1)}

\item{exportar}{Se TRUE, exporta para Excel}

\item{caminho}{Caminho para salvar o arquivo}
}
\value{
Data frame com a matriz
}
\description{
Gera matriz de planejamento fatorial completo
}
