% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_ffd.R
\name{metricas_ffd}
\alias{metricas_ffd}
\title{Calcula métricas do modelo fatorial}
\usage{
metricas_ffd(modelo)
}
\arguments{
\item{modelo}{Objeto retornado por fit_ffd()}
}
\value{
Data frame com métricas do modelo ajustado
}
\description{
Calcula métricas do modelo fatorial
}
\examples{
# metricas_ffd(fit)
}
