% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otimo_ffd.R
\name{otimo_ffd}
\alias{otimo_ffd}
\title{Ponto ótimo previsto para FFD}
\usage{
otimo_ffd(modelo, objetivo = c("min", "max"))
}
\arguments{
\item{modelo}{Objeto retornado por fit_ffd().}

\item{objetivo}{"min" para minimizar a resposta ou "max" para maximizar.}
}
\value{
Lista com ponto ótimo previsto, resposta estimada, convergência e objetivo.
}
\description{
Ponto ótimo previsto para FFD
}
