% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_ffd.R
\name{pareto_ffd}
\alias{pareto_ffd}
\title{Gráfico de Pareto dos efeitos padronizados no FFD}
\usage{
pareto_ffd(modelo, alpha = 0.05)
}
\arguments{
\item{modelo}{Objeto retornado por fit_ffd().}

\item{alpha}{Nível de significância.}
}
\value{
Invisivelmente, a tabela usada no gráfico.
}
\description{
Gráfico de Pareto dos efeitos padronizados no FFD
}
