% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponto_estacionario_ffd.R
\name{ponto_estacionario_ffd}
\alias{ponto_estacionario_ffd}
\title{Calcular ponto estacionário do modelo FFD}
\usage{
ponto_estacionario_ffd(modelo)
}
\arguments{
\item{modelo}{Objeto retornado por fit_ffd().}
}
\value{
Lista com ponto estacionário, classificação, autovalores,
matriz B, resposta estimada e status.
}
\description{
Calcular ponto estacionário do modelo FFD
}
