% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.avaliar_ponto_estacionario_ffd.R
\name{print.avaliar_ponto_estacionario_ffd}
\alias{print.avaliar_ponto_estacionario_ffd}
\title{Imprime avaliação do ponto estacionário FFD}
\usage{
\method{print}{avaliar_ponto_estacionario_ffd}(x, ...)
}
\arguments{
\item{x}{Objeto da classe avaliar_ponto_estacionario_ffd.}

\item{...}{Argumentos adicionais.}
}
\value{
Invisivelmente, o próprio objeto.
}
\description{
Imprime avaliação do ponto estacionário FFD
}
