% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.otimo_ffd.R
\name{print.otimo_ffd}
\alias{print.otimo_ffd}
\title{Imprime resultado do ponto ótimo FFD}
\usage{
\method{print}{otimo_ffd}(x, ...)
}
\arguments{
\item{x}{Objeto da classe otimo_ffd.}

\item{...}{Argumentos adicionais.}
}
\value{
Invisivelmente, o próprio objeto.
}
\description{
Imprime resultado do ponto ótimo FFD
}
