% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ponto_estacionario_ffd.R
\name{print.ponto_estacionario_ffd}
\alias{print.ponto_estacionario_ffd}
\title{Imprime ponto estacionário FFD}
\usage{
\method{print}{ponto_estacionario_ffd}(x, ...)
}
\arguments{
\item{x}{Objeto da classe ponto_estacionario_ffd.}

\item{...}{Argumentos adicionais.}
}
\value{
Invisivelmente, o próprio objeto.
}
\description{
Imprime ponto estacionário FFD
}
