% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clipboard_ffd.R
\name{read_clipboard_ffd}
\alias{read_clipboard_ffd}
\title{Lê dados copiados do Excel via clipboard}
\usage{
read_clipboard_ffd()
}
\value{
Um data frame com os dados copiados.
}
\description{
Lê uma tabela copiada diretamente do Excel para dentro do R.
}
\examples{
# Copie uma tabela no Excel e depois execute:
# dados <- read_clipboard_ffd()
}
