% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superficie_ffd.R
\name{superficie_ffd}
\alias{superficie_ffd}
\title{Gráfico de superfície de resposta para FFD}
\usage{
superficie_ffd(modelo, x1, x2, n = 45)
}
\arguments{
\item{modelo}{Objeto retornado por fit_ffd().}

\item{x1}{Nome do primeiro fator.}

\item{x2}{Nome do segundo fator.}

\item{n}{Número de pontos da grade.}
}
\value{
Invisivelmente, uma lista com grade e matriz de predições.
}
\description{
Gráfico de superfície de resposta para FFD
}
